trading/ai-trading/ollama-bot-files/requirements.txt
discord.py>=2.3.0
ollama>=0.1.0
requests>=2.31.0
